Add skeleton for new demo
