+++ Demo/Assets.xcassets/AppIcon.appiconset/Contents.json
