--- Demo/Common/State/CounterState.swift
@@ -1,13 +1,0 @@
-//
-//  CounterState.swift
-//  Katana
-//
-//  Copyright © 2016 Bending Spoons.
-//  Distributed under the MIT License.
-//  See the LICENSE file for more information.
-
-import Katana
-
-struct CounterState: State {
-  var counter: Int = 0
-}
