+++ Demo/Logic/AppDependenciesContainer.swift
@@ -1,0 +1,15 @@
+//
+//  AppDependenciesContainer.swift
+//  Demo
+//
+//  Created by Mauro Bolis on 19/02/2018.
+//
+
+import Foundation
+import Katana
+
+final class AppDependenciesContainer: SideEffectDependencyContainer {
+  init(dispatch: @escaping StoreDispatch, getState: @escaping () -> State) {
+    
+  }
+}
