+++ Demo/Logic/AppState.swift
@@ -1,0 +1,13 @@
+//
+//  AppState.swift
+//  Demo
+//
+//  Created by Mauro Bolis on 19/02/2018.
+//
+
+import Foundation
+import Katana
+
+struct AppState: State {
+  
+}
