+++ Demo/ViewController.swift
@@ -1,0 +1,18 @@
+//
+//  ViewController.swift
+//  Demo
+//
+//  Created by Mauro Bolis on 19/02/2018.
+//  Copyright © 2018 Mauro Bolis. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    self.view.backgroundColor = .red
+  }
+}
+
